Revert "added knife sound and arrow to button"
