+++ comps/Continue.js
@@ -49,7 +49,7 @@ export default function Continue() {
     page = 0;
   }
 
-  if (r.asPath === "/animalselection") {
+  if (r.asPath === "/") {
     return <ButtonCont as={motion.div} className="tipCont" initial="onLoad" animate="visible" variants={{
       onLoad: {
         scale: .4,
@@ -64,7 +64,7 @@ export default function Continue() {
         () => {
           {play("/sounds/ping.mp3")};
           r.push("/weapons");
-      }}>Continue ➜</button>
+      }}>Continue</button>
     </ButtonCont>
   } else if (r.asPath === "/weapons") {
       if (a === "Chicken") {
@@ -82,7 +82,7 @@ export default function Continue() {
             () => {
               {play("/sounds/ping.mp3")};
               r.push("/chicken");
-          }}>Continue ➜</button>
+          }}>Continue</button>
         </ButtonCont>
       } else if (a === "Cow") {
         return <ButtonCont as={motion.div} className="tipCont" initial="onLoad" animate="visible" variants={{
@@ -99,7 +99,7 @@ export default function Continue() {
               () => {
                 {play("/sounds/ping.mp3")};
                 r.push("/cow")
-            }}>Continue ➜</button>
+            }}>Continue</button>
           </ButtonCont>
       } else {
         return <ButtonCont as={motion.div} className="tipCont" initial="onLoad" animate="visible" variants={{
@@ -116,7 +116,7 @@ export default function Continue() {
             () => {
               { play("/sounds/ping.mp3") };
               r.push("pig")
-          }}>Continue ➜</button>
+          }}>Continue</button>
         </ButtonCont>
         }
   } else if (r.pathname === "/chicken") {
@@ -155,7 +155,7 @@ export default function Continue() {
             query: {
               page: Number(page) + 1
             }
-          })}}>Continue ➜</button>
+          })}}>Continue</button>
       </ButtonCont>
   } else if (r.pathname === "/cow") {
     return <ButtonCont as={motion.div} className="tipCont" initial="onLoad" animate="visible" variants={{
@@ -176,7 +176,7 @@ export default function Continue() {
             page: Number(page) + 1
           }
         })
-      }}>Continue ➜</button>
+      }}>Continue</button>
     </ButtonCont>
   } else if (r.pathname === "/pig") {
     return <ButtonCont as={motion.div} className="tipCont" initial="onLoad" animate="visible" variants={{
@@ -197,7 +197,7 @@ export default function Continue() {
             page: Number(page) + 1
           }
         })
-      }}>Continue ➜</button>
+      }}>Continue</button>
     </ButtonCont>
   }
 }
